+++ src/engine/components/index.ts
@@ -44,6 +44,7 @@ export {
   renderDivider,
   renderTimeline,
 } from "./FeedbackComponents";
+export { renderLayoutGroup, renderLabelledGroup } from "./LayoutComponents";
 export { renderChart } from "./ChartComponent";
 export { renderFiltersPanel } from "./FiltersPanelComponent";
 export { renderActiveFiltersBar } from "./ActiveFiltersBarComponent";